+++ mod/src/test/java/org/shsts/tinactory/gametest/StructureHelper.java
@@ -1,4 +1,4 @@
-package org.shsts.tinactory.test;
+package org.shsts.tinactory.gametest;
 
 import javax.annotation.ParametersAreNonnullByDefault;
 import net.minecraft.MethodsReturnNonnullByDefault;
