+++ mod/src/test/java/org/shsts/tinactory/gametest/mixin/MixinStructureUtils.java
@@ -1,11 +1,11 @@
-package org.shsts.tinactory.test.mixin;
+package org.shsts.tinactory.gametest.mixin;
 
 import javax.annotation.ParametersAreNonnullByDefault;
 import net.minecraft.MethodsReturnNonnullByDefault;
 import net.minecraft.gametest.framework.StructureUtils;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.world.level.levelgen.structure.templatesystem.StructureTemplate;
-import org.shsts.tinactory.test.StructureHelper;
+import org.shsts.tinactory.gametest.StructureHelper;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.injection.At;
 import org.spongepowered.asm.mixin.injection.Inject;
