+++ mod/src/test/resources/mixins.tinactory.test.json
@@ -1,6 +1,6 @@
 {
   "required": false,
-  "package": "org.shsts.tinactory.test.mixin",
+  "package": "org.shsts.tinactory.gametest.mixin",
   "compatibilityLevel": "JAVA_17",
   "minVersion": "0.8",
   "mixins": [
